Add some initial gui

+++ src/main/java/com/syrnik/repository/central/BranchRepository.java
@@ -4,6 +4,6 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.syrnik.model.central.Branch;
 
-public interface BranchRepository extends JpaRepository<Branch, Long> {
+public interface BranchRepository extends JpaRepository<Branch, Integer> {
     Optional<Branch> findByName(String name);
 }
